move note fetching back to index
